+++ src/interfaces/middleware-handler.ts
@@ -1,6 +1,5 @@
 import { CustomRequest } from '../lib/custom-request';
 import { CustomResponse } from '../lib/custom-response';
-import { NextResponse } from 'next/server';
 import { NextHandler } from './util-types';
 
 export interface IMiddlewareHandler<
@@ -14,5 +13,5 @@ export interface IMiddlewareHandler<
     req: CustomRequest<TParams, TQuery, TPayload, TSession>,
     res: CustomResponse<TResponseData>,
     next: NextHandler,
-  ): Promise<NextResponse<TResponseData> | void>;
+  ): Promise<CustomResponse<TResponseData> | void>;
 }