+++ src/lib/custom-response.ts
@@ -2,6 +2,14 @@
 import { NextResponse } from 'next/server';
 import { BodyInit, CookieOptions, ResponseInit } from '../interfaces';
 
+/**
+ * TODO: Move this to a shared utility file and add tests
+ * The following function is used to capitalize the first letter of a string.
+ */
+function capitalizeFirstLetter(str: string) {
+  return str.charAt(0).toUpperCase() + str.slice(1);
+}
+
 export class CustomResponse<TResponseData = unknown> extends NextResponse {
   private noBodyStatusCodes = [204, 205, 304];
   /**
@@ -110,6 +118,8 @@ export class CustomResponse<TResponseData = unknown> extends NextResponse {
       }
       if (options.path) {
         cookieString += `; Path=${options.path}`;
+      } else {
+        cookieString += `; Path=/`;
       }
       if (options.domain) {
         cookieString += `; Domain=${options.domain}`;
@@ -118,8 +128,14 @@ export class CustomResponse<TResponseData = unknown> extends NextResponse {
         cookieString += `; Secure`;
       }
       if (options.sameSite) {
-        cookieString += `; SameSite=${options.sameSite}`;
+        if (typeof options.sameSite === 'boolean') {
+          cookieString += `; SameSite=Strict`;
+        } else {
+          cookieString += `; SameSite=${capitalizeFirstLetter(options.sameSite)}`;
+        }
       }
+    } else {
+      cookieString += '; Path=/';
     }
 
     this.setHeader('Set-Cookie', cookieString);
@@ -127,10 +143,72 @@ export class CustomResponse<TResponseData = unknown> extends NextResponse {
   }
 
   /**
+   * The following method is used to clear a cookie in the response header.
+   * @param name The name of the cookie to clear.
+   * @param options The options for the cookie.
+   * @returns The current instance of the CustomResponse object.
+   */
+  clearCookie(name: string, options?: CookieOptions): this {
+    const opts = {
+      ...options,
+      expires: new Date(1), // Set the cookie expiration to a past date
+      path: options?.path || '/', // Default path to '/'
+      maxAge: undefined, // Ensure maxAge is cleared
+    };
+
+    // Call the existing cookie method with an empty value to clear it
+    return this.cookie(name, '', opts);
+  }
+
+  /**
+   * The following method is used to redirect the response to a different URL.
+   * It defaults to a 302 status code unless otherwise specified.
+   * @param url The URL to redirect to.
+   * @param statusCode The status code to use for the redirect.
+   * @returns CustomResponse
+   *
+   * @example
+   * response.redirect('https://example.com/new-url');
+   * response.redirect(301, 'https://example.com/new-url');
+   */
+  redirect(url: string): CustomResponse<unknown>;
+  redirect(statusCode: number, url: string): CustomResponse<unknown>;
+
+  redirect(statusCodeOrUrl: number | string, maybeUrl?: string) {
+    let url: string;
+    let statusCode = 302; // Default to 302
+
+    if (typeof statusCodeOrUrl === 'string') {
+      url = statusCodeOrUrl;
+    } else {
+      statusCode = statusCodeOrUrl;
+      url = maybeUrl!;
+    }
+
+    // Validate that the status code is a valid 3xx code for redirects
+    if (statusCode < 300 || statusCode >= 400) {
+      statusCode = 302; // Default to 302 if invalid
+    }
+
+    // Set the "Location" header for redirection
+    this.setHeader('Location', url);
+
+    // Use Next.js's NextResponse.redirect method for handling redirects
+    const response = NextResponse.redirect(url, statusCode);
+
+    // Set headers on the response
+    this.headers.forEach((value, key) => {
+      response.headers.set(key, value);
+    });
+
+    return response;
+  }
+
+  /**
    * The following method is used to send the response.
    * It defaults to a 200 status code unless otherwise specified.
    * @param body The body of the response.
-   * @returns A NextResponse object.
+   * @returns A CustomResponse object.
    */
   send(body?: TResponseData) {
     let response: NextResponse;
@@ -154,12 +232,12 @@ export class CustomResponse<TResponseData = unknown> extends NextResponse {
     });
 
     // TODO: Search for a better way to handle this
-    return response as NextResponse<TResponseData>;
+    return response as CustomResponse<TResponseData>;
   }
 
   /**
    * The following method sends a response with no body.
-   * @returns A NextResponse object.
+   * @returns A CustomResponse object.
    */
   end() {
     // Finalize the response without a body